pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "img2vid"
version = "0.1.0"
description = "根据多图片和字幕生成配音视频的工具"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "edge-tts>=6.1.0",
    "click>=8.0",
    "pillow>=10.0",
]

[project.scripts]
img2vid = "img2vid.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
